[fix] delete someone's message

+++ chat_delete.php
@@ -1,11 +1,15 @@
 <?php
 require_once 'system.php';
 
+auth_only();
+
 $id = $_GET['id'];
 
-mysqli_query($conn, "delete from chat where id = $id");
+$result = mysqli_query($conn, "select * from chat where id = $id");
+$post = mysqli_fetch_assoc($result);
 
-header('location: chat.php');
-exit;
+if ($post['user_id'] == $auth['id']) {
+    mysqli_query($conn, "delete from chat where id = $id");
+}
 
-?>
+redirect('chat.php');
